# schemas/ripd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RIP Agent Server",
    "description": "Agent session control plane (HTTP/SSE).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/sessions": {
      "post": {
        "operationId": "create_session",
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{id}/cancel": {
      "post": {
        "operationId": "cancel_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session canceled"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/sessions/{id}/events": {
      "get": {
        "operationId": "stream_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of event frames"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/sessions/{id}/input": {
      "post": {
        "operationId": "send_input",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Input accepted"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InputPayload": {
        "type": "object",
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "type": "string"
          }
        }
      },
      "SessionCreated": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      }
    }
  }
}